Add location.href to id URI for Annotator JSON
Annotator JSON does not (by default) store
the information about its storage location
--which, consequently, is not on a server by
default, but rather only in memory in the
browser. As such, this is a wild guess that
the annotation is being served from its
storage location.

I.e. this needs fixing because it MAY be right
but it probably is not...